restructure CI pipeline
