+++ src/lambda/ofgemCheck/indexCheckScraperOutput.ts
