+++ src/lambda/ofgemCheck/indexLogOfgemCheckOutputs.ts
@@ -1,6 +1,6 @@
 //top level index file for lambda
 
-import handler from "../ofgemCheck/getOfgemCheckInputs.js";
+import handler from "../../ofgemCheck/logOfgemCheckOutputs.js";
 
 module.exports = {
     handler
